Fix partition removal methods
There's no safe way for users to use the libparted API as it was written,
so I'm changing the API here to make it safer. Basically, rather than
having to get a mutable pointer to a partition from within a disk and then
reusing the disk mutably alongside the mutable child reference, this will
let the programmer pass the partition number to be deleted, and do that
internally.